Updated for MSigDB v6.1, removed Java dependencies
Modified to use GP-native file download facilities for GMT files instead
of the previous Java components.  This is possible because of the MSigDB
v6.1 open-access release.
Partially updated toward using the r.package.info method for installing
R dependencies.  This is not complete but is usable as a Beta on our
public server.
Added CLI module build support (Ant).
Copyright updates.